+++ chip-service/src/main/java/com/rljj/chipservice/domain/chipexchange/dto/ChipExchangeCreateRequest.java
@@ -1,14 +1,18 @@
-package com.rljj.chipservice.domain.exchange.dto;
+package com.rljj.chipservice.domain.chipexchange.dto;
 
 import com.rljj.switchswitchentity.chip.chipexchange.ChipExchangeStatus;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ChipExchangeStatusUpdateRequest {
-    private Long chipExchangeId;
+public class ChipExchangeCreateRequest {
+    private Long chipPostId;
+    private Long chipInfoId;
+    private String content;
     private ChipExchangeStatus status;
 }
