+++ chip-service/src/main/java/com/rljj/chipservice/domain/chipexchange/dto/ChipExchangeStatusUpdateRequest.java
@@ -1,18 +1,14 @@
-package com.rljj.chipservice.domain.exchange.dto;
+package com.rljj.chipservice.domain.chipexchange.dto;
 
 import com.rljj.switchswitchentity.chip.chipexchange.ChipExchangeStatus;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ChipExchangeCreateRequest {
-    private Long chipPostId;
-    private Long chipInfoId;
-    private String content;
+public class ChipExchangeStatusUpdateRequest {
+    private Long chipExchangeId;
     private ChipExchangeStatus status;
 }
