+++ chip-service/src/main/java/com/rljj/chipservice/domain/chipexchange/service/ChipExchangeServiceImpl.java
@@ -1,9 +1,9 @@
-package com.rljj.chipservice.domain.exchange.service;
+package com.rljj.chipservice.domain.chipexchange.service;
 
-import com.rljj.chipservice.domain.exchange.dto.ChipExchangeCreateRequest;
-import com.rljj.chipservice.domain.exchange.dto.ChipExchangeResponse;
-import com.rljj.chipservice.domain.exchange.dto.ChipExchangeStatusUpdateRequest;
-import com.rljj.chipservice.domain.exchange.repository.ChipExchangeRepository;
+import com.rljj.chipservice.domain.chipexchange.dto.ChipExchangeCreateRequest;
+import com.rljj.chipservice.domain.chipexchange.dto.ChipExchangeResponse;
+import com.rljj.chipservice.domain.chipexchange.dto.ChipExchangeStatusUpdateRequest;
+import com.rljj.chipservice.domain.chipexchange.repository.ChipExchangeRepository;
 import com.rljj.switchswitchcommon.exception.NotFoundException;
 import com.rljj.switchswitchentity.chip.chipexchange.ChipExchange;
 import com.rljj.switchswitchentity.chip.chipexchange.ChipExchangeStatus;
